Create SignIn.java
added SignIn.java, acitvity_signin.xml, edittext_border.xml (solely for the appearance of the EditText on the Chat activity), sign_out_button.xml (solely for the appearance of the sign out button on the SignIn activity), as well as google services dependencies and configuration file. There's some code in SignIn.java that has to do with signing out and disconnecting one's account, which I think we should make a button for on all activities.